move intersect code
