+++ app/helpers/Intersect.js
@@ -6,6 +6,10 @@
 
 import { GetStoreData, SetStoreData } from '../helpers/General';
 import { LOCATION_DATA, CROSSED_PATHS } from '../constants/storage';
+import PushNotification from 'react-native-push-notification';
+
+import { isPlatformiOS } from './../Util';
+import languages from '../locales/languages';
 
 export async function IntersectSet(concernLocationArray, completion) {
   GetStoreData(LOCATION_DATA).then(locationArrayString => {
@@ -105,12 +109,12 @@ export async function IntersectSet(concernLocationArray, completion) {
 /**
  * Function to determine if two location points are "nearby".
  * Uses shortcuts when possible, then the exact calculation.
- * 
+ *
  * @param {number} lat1 - location 1 latitude
  * @param {number} lon1 - location 1 longitude
  * @param {number} lat2 - location 2 latitude
  * @param {number} lon2 - location 2 longitude
- * @return {boolean} true if the two locations meet the criteria for nearby 
+ * @return {boolean} true if the two locations meet the criteria for nearby
  */
 export function isLocationsNearby(lat1, lon1, lat2, lon2) {
   let nearbyDistance = 20; // in meters, anything closer is "nearby"
@@ -204,4 +208,92 @@ function binarySearchForTime(array, targetTime) {
     }
   }
   return -i - 1;
+}
+
+export function checkIntersect() {
+  // This function is called once every 12 hours.  It should do several things:
+
+  console.log(
+    'Intersect tick entering on',
+    isPlatformiOS() ? 'iOS' : 'Android',
+  );
+  // this.findNewAuthorities(); NOT IMPLEMENTED YET
+
+  // Get the user's health authorities
+  GetStoreData('AUTHORITY_SOURCE_SETTINGS')
+    .then(authority_list => {
+      if (!authority_list) {
+        // DEBUG: Force a test list
+        // authority_list = [
+        //  {
+        //    name: 'Platte County Health',
+        //    url:
+        //      'https://raw.githack.com/tripleblindmarket/safe-places/develop/examples/safe-paths.json',
+        //  },
+        //];
+        console.log('No authorities', authority_list);
+        return;
+      }
+
+      let name_news = [];
+      if (authority_list) {
+        // Pull down data from all the registered health authorities
+        authority_list = JSON.parse(authority_list);
+        for (const authority of authority_list) {
+          console.log('[auth]', authority);
+          fetch(authority.url)
+            .then(response => response.json())
+            .then(responseJson => {
+              // Example response =
+              // { "authority_name":  "Steve's Fake Testing Organization",
+              //   "publish_date_utc": "1584924583",
+              //   "info_website": "https://www.who.int/emergencies/diseases/novel-coronavirus-2019",
+              //   "concern_points":
+              //    [
+              //      { "time": 123, "latitude": 12.34, "longitude": 12.34},
+              //      { "time": 456, "latitude": 12.34, "longitude": 12.34}
+              //    ]
+              // }
+
+              // Update cache of info about the authority
+              // TODO: Add an "info_newsflash" UTC timestamp and popup a
+              //       notification if that changes, i.e. if there is a newsflash?
+              name_news.push({
+                name: responseJson.authority_name,
+                news_url: responseJson.info_website,
+              });
+
+              // TODO: Look at "publish_date_utc".  We should notify users if
+              //       their authority is no longer functioning.)
+
+              IntersectSet(responseJson.concern_points, dayBin => {
+                if (dayBin !== null && dayBin.reduce((a, b) => a + b, 0) > 0) {
+                  PushNotification.localNotification({
+                    title: languages.t('label.push_at_risk_title'),
+                    message: languages.t('label.push_at_risk_message'),
+                  });
+                }
+              });
+            });
+
+          SetStoreData('AUTHORITY_NEWS', name_news)
+            .then(() => {
+              // TODO: Anything after this saves?  Background caching of
+              //       news to make it snappy?  Could be a problem in some
+              //       locales with high data costs.
+            })
+            .catch(error =>
+              console.log('Failed to save authority/news URL list', error),
+            );
+        }
+        let nowUTC = new Date().toISOString();
+        let unixtimeUTC = Date.parse(nowUTC);
+        // Last checked key is not being used atm. TODO check this to update periodically instead of every foreground activity
+        SetStoreData('LAST_CHECKED', unixtimeUTC);
+      } else {
+        console.log('No authority list');
+        return;
+      }
+    })
+    .catch(error => console.log('Failed to load authority list', error));
 }
